# Refactor the internals of antialiasing / backbuffer on WebGL device